requirements.txt
django
requests
django-cors-headers
djangorestframework
pycryptodome
cryptography
